+++ CHANGES.md
@@ -1,5 +1,6 @@
 # 0.4.0-alpha.1
 ## 🌟 Features
+- You can now toggle todo, bullet and numbered lists directly from the floating menu
 - Word count can now be always shown in the bottom-right corner of the window. (You can enable or disable it in settings)
 
 ## ✨ Improvements and fixes